backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
Werkzeug==3.0.1
groq==0.11.0
httpx==0.27.2
requests==2.31.0
